Updated quasideuteron.ipynb
Updating figure functions to save to figures/quasideuteron. So far, we can plot the ratio of pn/d momentum distributions, the Levinger constant with respect to A, print a table of Levinger constants, and look at the asymptotic relation between proton and pair momentum distributions. In all cases, we compare to experiment or AV18 VMC calculations when possible.